pyproject.toml
[project]
name = "shorts-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "ipykernel>=7.1.0",
    "langchain>=1.2.10",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.9",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
    "rich>=14.3.3",
]
